Move leaflet-layer to native dir
