+++ test/spec/api/v4/client.spec.js
@@ -1,6 +1,6 @@
 var L = require('leaflet');
 var carto = require('../../../../src/api/v4');
-var LeafletLayer = require('../../../../src/api/v4/leaflet-layer');
+var LeafletLayer = require('../../../../src/api/v4/native/leaflet-layer');
 
 describe('api/v4/client', function () {
   var client;
